separate logic from DOM related stuff
